Factories to detect and flag duplicated and multiple values of a variable.

These functions create other functions to detect and flag multiple values of a specific variable. Their goal is to create expressive predicates (detect_*()) that can be used in, for example, if() statements; and expressive functions to create messages when a condition is true (flag_*()).

detect_multiple_f(name)

detect_duplicated_f(name)

flag_multiple_f(name, cond = warning)

flag_duplicated_f(name, cond = warning)

Arguments

name
String; the name of a variable of .data.

cond
Symbol; the bare name of a function that outputs a condition: e.g. warning, stop, message, rlang::warn, rlang::abort, rlang::inform.

Value

A function.

Arguments to the resulting function

.data: A dataframe.

msg: String; an optional custom message.

See also

Examples

# DETECT ------------------------------------------------------------------
dfm <- data.frame(CensusID = c(1, 2, NA))
censusid_has_multiple_values <- detect_multiple_f("censusid")
if (censusid_has_multiple_values(dfm)) "Hello world"

#> [1] "Hello world"

censusid_has_multiple_values(data.frame(CensusID = c(1, 1, NA)))

#> [1] FALSE

# Insensitive to upper/lowercase
multiple_censusid <- detect_multiple_f("censusid")
multiple_censusid(data.frame(censusid = c(1, 2, NA)))

#> [1] TRUE

multiple_censusid <- detect_multiple_f("CENSUSID")
multiple_censusid(data.frame(censusid = c(1, 2, NA)))

#> [1] TRUE

# FLAG --------------------------------------------------------------------
.df <- data.frame(a = 1:3, b = 1, stringsAsFactors = FALSE)
flag_multiple_f("a")(.df)

#> Warning: a: Multiple values were detected.

flag_multiple_f("a", message)(.df, "Hello world.")

#> Hello world.

# Silent
flag_duplicated_f("a")(.df)
.df2 <- data.frame(a = c(1, 1, 1), stringsAsFactors = FALSE)
flag_duplicated_f("a")(.df2)

#> Warning: a: Duplicated values were detected.

# NOT RUN {
# Dealing with grouped data
if (!requireNamespace("dplyr")) {
  library(dplyr)
  # `b` is single within groups but multiple accross entire dataset
  .df <- tibble(a = c(1, 1, 2, 2), b = c(1, 1, 2, 2))

  by_x <- group_by(.df, a)
  # Works accross entire dataset -- expect warning
  flag_if_multiple_b <- flag_multiple_f("b")
  flag_if_multiple_b(by_x)
  # Works within groups -- expect silent
  do(by_x, (flag_if_multiple_b(.)))
  # Same, with interface similar to lapply() and purrr::map()
  fgeo.tool::by_group(by_x, flag_if_multiple_b)
  # Also consider tidyr::nest() + dplyr::mutate() + dpyr::map())
}
# }