Create a vector or dataframe with case insensitive name matching.

Create a vector or dataframe with case insensitive name matching.

insensitive(x)

Arguments

x
vector or dataframe to modify.

Acknowledgment

Thanks to Neil Richardson for recommending this function (https://twitter.com/enpiar).

See also

Examples

x <- c("abc" = 1, "def" = 2)
x["ABC"]

#> <NA> 
#>   NA 

y <- insensitive(x)
y["ABC"]

#> abc 
#>   1 

y[["ABC"]]

#> [1] 1

y$ABC

#> [1] 1

vft <- data.frame(TreeID = 1)
insensitive(vft)[["TreeID"]]

#> [1] 1

insensitive(vft)[["treeid"]]

#> [1] 1

insensitive(vft)["TreeID"]

#>   treeid
#> 1      1

insensitive(vft)["treeid"]

#>   treeid
#> 1      1

insensitive(vft)$TreeID

#> [1] 1

insensitive(vft)$treeid

#> [1] 1

# Works
transform(insensitive(vft), new = treeid)

#>   treeid new
#> 1      1   1

# dplyr::mutate(insensitive(vft), new = treeid)
# Fails
# transform(insensitive(vft), new = TreeID)
# dplyr::mutate(insensitive(vft), new = TreeID)